Split local H5 contracts into themed modules
