+++ tests/unit/calibration/fixtures/__init__.py
@@ -1,0 +1,5 @@
+"""Shared test helpers for unit calibration tests.
+
+Fixture modules in this package hold setup code and reusable test-only
+helpers so individual test files stay focused on assertions.
+"""
